[#89] Reorganise Cardano.Crypto.Signing module tree
