+++ binary/stack.yaml
@@ -6,7 +6,7 @@ packages:
 
 extra-deps:
   - git: https://github.com/input-output-hk/cardano-prelude
-    commit: 0889d18d3581026598b9268637475525064b76fe
+    commit: 3bb05f785ef6427a87af43517c561713e8314c94
     subdirs:
       - .
       - test
