+++ crypto/cardano-crypto-wrapper.cabal
@@ -1,4 +1,5 @@
 name:                cardano-crypto-wrapper
+
 version:             1.3.0
 synopsis:            Cryptographic primitives used in the Cardano project
 description:         Cryptographic primitives used in the Cardano project
@@ -31,17 +32,31 @@ library
                        Cardano.Crypto.Signing
                        Cardano.Crypto.Signing.Redeem
                        Cardano.Crypto.Signing.Safe
-                       Cardano.Crypto.Signing.Signing
 
   other-modules:
                        Cardano.Crypto.Scrypt
 
-                       Cardano.Crypto.Signing.Check
                        Cardano.Crypto.Signing.Tag
-                       Cardano.Crypto.Signing.Types.Redeem
-                       Cardano.Crypto.Signing.Types.Safe
-                       Cardano.Crypto.Signing.Types.Signing
-                       Cardano.Crypto.Signing.Types.Tag
+
+                       Cardano.Crypto.Signing.KeyGen
+                       Cardano.Crypto.Signing.PublicKey
+                       Cardano.Crypto.Signing.SecretKey
+                       Cardano.Crypto.Signing.Signature
+
+                       Cardano.Crypto.Signing.Proxy
+                       Cardano.Crypto.Signing.Proxy.Cert
+                       Cardano.Crypto.Signing.Proxy.SecretKey
+                       Cardano.Crypto.Signing.Proxy.Signature
+
+                       Cardano.Crypto.Signing.Redeem.KeyGen
+                       Cardano.Crypto.Signing.Redeem.PublicKey
+                       Cardano.Crypto.Signing.Redeem.SecretKey
+                       Cardano.Crypto.Signing.Redeem.Signature
+
+                       Cardano.Crypto.Signing.Safe.EncryptedSecretKey
+                       Cardano.Crypto.Signing.Safe.KeyGen
+                       Cardano.Crypto.Signing.Safe.PassPhrase
+                       Cardano.Crypto.Signing.Safe.SafeSigner
 
   build-depends:       base
                      , aeson
