+++ crypto/src/Cardano/Crypto/Signing.hs
@@ -3,8 +3,11 @@ module Cardano.Crypto.Signing
   )
 where
 
-import Cardano.Crypto.Signing.Check as X
+import Cardano.Crypto.Signing.KeyGen as X
+import Cardano.Crypto.Signing.Proxy as X
+import Cardano.Crypto.Signing.PublicKey as X
 import Cardano.Crypto.Signing.Redeem as X
 import Cardano.Crypto.Signing.Safe as X
-import Cardano.Crypto.Signing.Signing as X
+import Cardano.Crypto.Signing.SecretKey as X
+import Cardano.Crypto.Signing.Signature as X
 import Cardano.Crypto.Signing.Tag as X
