+++ crypto/src/Cardano/Crypto/Signing/Proxy.hs
@@ -1,0 +1,8 @@
+module Cardano.Crypto.Signing.Proxy
+  ( module X
+  )
+where
+
+import Cardano.Crypto.Signing.Proxy.Cert as X
+import Cardano.Crypto.Signing.Proxy.SecretKey as X
+import Cardano.Crypto.Signing.Proxy.Signature as X
