+++ crypto/src/Cardano/Crypto/Signing/Redeem/PublicKey.hs
@@ -9,22 +9,18 @@
 {-# LANGUAGE TemplateHaskell            #-}
 {-# LANGUAGE UndecidableInstances       #-}
 
-module Cardano.Crypto.Signing.Types.Redeem
-  ( RedeemSecretKey(..)
-  , RedeemPublicKey(..)
-  , RedeemSignature(..)
+module Cardano.Crypto.Signing.Redeem.PublicKey
+  ( RedeemPublicKey(..)
   , redeemPkB64F
   , redeemPkB64UrlF
   , redeemPkB64ShortF
   , fromAvvmPk
   , redeemPkBuild
-  , redeemToPublic
   )
 where
 
 import Cardano.Prelude
 
-import Control.Monad.Except (MonadError)
 import Crypto.Error (CryptoFailable(..))
 import qualified Crypto.PubKey.Ed25519 as Ed25519
 import Data.Aeson
@@ -33,7 +29,7 @@ import Data.Aeson
   , ToJSONKey(..)
   , ToJSONKeyFunction(..)
   )
-import qualified Data.Aeson.Encoding as A (text)
+import qualified Data.Aeson.Encoding as A
 import Data.Aeson.TH (defaultOptions, deriveJSON)
 import qualified Data.ByteArray as BA
 import qualified Data.ByteString as BS
@@ -43,16 +39,12 @@ import qualified Data.ByteString.Char8 as Char8
 import qualified Data.Text as T
 import Formatting
   (Format, bprint, build, fitLeft, formatToString, later, sformat, stext, (%.))
-import qualified Formatting.Buildable as B (Buildable(..))
+import qualified Formatting.Buildable as B
 import Text.JSON.Canonical (FromObjectKey(..), JSValue(..), ToObjectKey(..))
 
 import Cardano.Binary.Class (Bi)
 import Cardano.Crypto.Orphans ()
 
-
---------------------------------------------------------------------------------
--- PK/SK and formatters
---------------------------------------------------------------------------------
 
 -- | Wrapper around 'Ed25519.PublicKey'.
 newtype RedeemPublicKey =
@@ -79,12 +71,10 @@ instance FromJSONKey RedeemPublicKey where
       . bimap (sformat build) pure
       . fromAvvmPk
 
--- | Wrapper around 'Ed25519.SecretKey'.
-newtype RedeemSecretKey =
-  RedeemSecretKey Ed25519.SecretKey
-  deriving (Eq, Ord, Show, Generic, NFData)
+deriving instance Bi RedeemPublicKey
 
-deriving instance Bi RedeemSecretKey
+instance B.Buildable RedeemPublicKey where
+  build = bprint ("redeem_pk:" . redeemPkB64F)
 
 fromPublicKeyToByteString :: Ed25519.PublicKey -> BS.ByteString
 fromPublicKeyToByteString = BA.convert
@@ -100,44 +90,6 @@ redeemPkB64UrlF = later $ \(RedeemPublicKey pk) ->
 
 redeemPkB64ShortF :: Format r (RedeemPublicKey -> r)
 redeemPkB64ShortF = fitLeft 8 %. redeemPkB64F
-
--- | Public key derivation function.
-redeemToPublic :: RedeemSecretKey -> RedeemPublicKey
-redeemToPublic (RedeemSecretKey k) = RedeemPublicKey (Ed25519.toPublic k)
-
-instance B.Buildable RedeemPublicKey where
-  build = bprint ("redeem_pk:" . redeemPkB64F)
-
-instance B.Buildable RedeemSecretKey where
-  build = bprint ("redeem_sec_of_pk:" . redeemPkB64F) . redeemToPublic
-
-deriving instance Bi RedeemPublicKey
-
-
---------------------------------------------------------------------------------
--- Redeem signatures
---------------------------------------------------------------------------------
-
--- | Wrapper around 'Ed25519.Signature'
-newtype RedeemSignature a =
-  RedeemSignature Ed25519.Signature
-  deriving (Eq, Ord, Show, Generic, NFData, Bi)
-
-instance B.Buildable (RedeemSignature a) where
-  build _ = "<redeem signature>"
-
-data AvvmPkError
-  = ApeAddressFormat Text
-  | ApeAddressLength Int
-  deriving (Show)
-
-instance B.Buildable AvvmPkError where
-  build = \case
-    ApeAddressFormat addrText ->
-      bprint ("Address " . stext . " is not base64(url) format") addrText
-    ApeAddressLength len -> bprint
-      ("Address length is " . build . ", expected 32, can't be redeeming pk")
-      len
 
 -- | Read the text into a redeeming public key. The key should be in
 --   AVVM format which is base64(url). This function must be inverse of
@@ -168,9 +120,17 @@ redeemPkBuild bs
       "Cardano.Crypto.Signing.Types.Redeem.hs consRedeemPk failed because "
       (T.pack $ show e)
 
--- These are *not* orphan instances, these types are defined in this file.
--- However these need to be defined here to avoid TemplateHaskell compile
--- phase errors.
+data AvvmPkError
+  = ApeAddressFormat Text
+  | ApeAddressLength Int
+  deriving (Show)
+
+instance B.Buildable AvvmPkError where
+  build = \case
+    ApeAddressFormat addrText ->
+      bprint ("Address " . stext . " is not base64(url) format") addrText
+    ApeAddressLength len -> bprint
+      ("Address length is " . build . ", expected 32, can't be redeeming pk")
+      len
 
 deriveJSON defaultOptions ''RedeemPublicKey
-deriveJSON defaultOptions ''RedeemSignature
