+++ crypto/src/Cardano/Crypto/Signing/Tag.hs
@@ -1,21 +1,61 @@
+{-# LANGUAGE DeriveGeneric     #-}
 {-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE OverloadedStrings #-}
 
 module Cardano.Crypto.Signing.Tag
-  ( signTag
-  , module Cardano.Crypto.Signing.Types.Tag
+  ( SignTag(..)
+  , signTag
   )
 where
 
 import Cardano.Prelude
 
+import Formatting (bprint, shown)
+import Formatting.Buildable (Buildable(..))
+
 import qualified Cardano.Binary.Class as Bi
 import Cardano.Crypto.ProtocolMagic (ProtocolMagic(..))
-import Cardano.Crypto.Signing.Types.Tag
 
+
+-- | To protect against replay attacks (i.e. when an attacker intercepts a
+--   signed piece of data and later sends it again), we add a tag to all data
+--   that we sign. This ensures that even if some bytestring can be deserialized
+--   into two different types of messages (A and B), the attacker can't take
+--   message A and send it as message B.
+--
+--   We also automatically add the network tag ('protocolMagic') whenever it
+--   makes sense, to ensure that things intended for testnet won't work for
+--   mainnet.
+data SignTag
+  = SignForTestingOnly
+  -- ^ Anything (to be used for testing only)
+  | SignTx
+  -- ^ Tx:               @TxSigData@
+  | SignRedeemTx
+  -- ^ Redeem tx:        @TxSigData@
+  | SignVssCert
+  -- ^ Vss certificate:  @(VssPublicKey, EpochIndex)@
+  | SignUSProposal
+  -- ^ Update proposal:  @UpdateProposalToSign@
+  | SignCommitment
+  -- ^ Commitment:       @(EpochIndex, Commitment)@
+  | SignUSVote
+  -- ^ US proposal vote: @(UpId, Bool)@
+  | SignMainBlock
+  -- ^ Main block:       @MainToSign@
+  | SignMainBlockHeavy
+  | SignProxySK
+  -- ^ Proxy key:        @ProxySecretKey@
+  deriving (Eq, Ord, Show, Generic)
+
+-- TODO: it would be nice if we couldn't use 'SignTag' with wrong
+-- types. Maybe something with GADTs and data families?
+
+instance Buildable SignTag where
+  build = bprint shown
 
 -- | Get magic bytes corresponding to a 'SignTag'. Guaranteed to be different
--- (and begin with a different byte) for different tags.
+--   (and begin with a different byte) for different tags.
 signTag :: ProtocolMagic -> SignTag -> ByteString
 signTag protocolMagic = \case
   SignForTestingOnly -> "\x00"
