+++ crypto/stack.yaml
@@ -6,10 +6,16 @@ packages:
 
 extra-deps:
   - git: https://github.com/input-output-hk/cardano-prelude
-    commit: 0889d18d3581026598b9268637475525064b76fe
+    commit: 3bb05f785ef6427a87af43517c561713e8314c94
     subdirs:
       - .
       - test
+  - git: https://github.com/well-typed/cborg # https://github.com/well-typed/cborg/pull/185
+    # This adds the capability to extract byte offsets in decoders
+    commit: 80fbe0ee5e67a5622e2cb9eaa9d8594a2214322d
+    subdirs:
+      - cborg
+
 
   - ../binary
   - ../binary/test
