--- crypto/test/golden/Signed
@@ -1,5 +1,0 @@
-00: 82f65840b28bf6b4c08892f5b82dc11e
-10: 9df08c0c4374a8f7e2b2558d42a83ed4
-20: 3ec2b6b76dabf1293a561971d772d3f9
-30: 6097f96f40a363b4201616202ef18fe5
-40: 2f27f806
